requirements.txt
PyMuPDF==1.25.5
requests==2.31.0
pillow
ollama